requirements.txt
pandas
scikit-learn
python-telegram-bot
psycopg2-binary
SQLAlchemy
python-dotenv